Fix race condition
phr_parse_* functions expect first parameter to point to the beginning of the data.

In our cases we passed pointer to *processed* data. In case if several responses
come from the same peer then the first response applies to both client responses
and it causes incorrect response handling and later crah due to assert.