Stream downloaded file to the client at the same time as we save it to cache
With the current implementation the client needs to wait until the
cached file is fully downloaded from the server and stored to local
filesystem. Once the file is cached at pacoloco only then it starts
serving it to the client.

This might take time, especially with a slow connection to the server.
And it causes timeout error at the client side.

Instead of waiting until the end of the download we stream the date to the
client while saving it to the cache.

Closes #5
Closes #6