Fix "too many SQL variables" sqlite restriction
There was an issue with sqlite3/GORM which happened when pacoloco tried saving the content of community.db file during the prefetching procedure: too many SQL variables.

This is due to the fact that GORM failed to save the file.db data into the db when a batch insertion is done. It tried to create a batch insertion of ~8.5k entries, which is too much for sqlite3, so I had split the data into <=2k entries. 2k is an arbitrary value, but I haven't found so far clear specifications on which limit should I set. There was no config in GORM to set it up automatically (CreateBatchSize could be relevant, but it seems that it is used only for table creation), so maybe we can open an issue about it in GORM.
Before this fix, pacoloco failed silently because db.Save() return value was discarded, as GORM documentation examples do.