Compile Docker image without CGO
As it required by go-sqlite3 dependency:
```
pacoloco    | prefetch.go:67: Binary was compiled with 'CGO_ENABLED=0', go-sqlite3 requires cgo to work. This is a stub
pacoloco exited with code 1
```

It switches the resulting docker image from a scratch base to alpine, which allows libraries that require CGO such as go-sqlite3 to work.

Closes #33